refactor(task-date-manager): move date filters from task-filter & remove redundant code
