+++ src/features/task-filter/domain/task-filters.ts
@@ -101,7 +101,9 @@ export const isInProgress = (
 	selectedAlias: string | null
 ): boolean => {
 	return (
-		!isCompleted(task) && !isCancelled(task) && !isSnoozed(task, taskMap, selectedAlias)
+		!isCompleted(task) &&
+		!isCancelled(task) &&
+		!isSnoozed(task, taskMap, selectedAlias)
 	);
 };
 
@@ -293,146 +295,4 @@ export const isBlankTask = (task: TaskItem): boolean => {
 	return txt.trim().length === 0;
 };
 
-/**
- * Schedule parsing: detect day-of-week schedules like:
- * - "🗓️ Sundays" / "🗓️ Sunday" (case-insensitive)
- * - "🗓 Sundays:" (with punctuation)
- * - "🗓️ Mon-Fri", "🗓️ Monday–Friday", "🗓️ Weekdays"
- * - "🗓️ Weekends" (Sat + Sun)
- * - "🗓️ Daily", "🗓️ Every day"
- * - Lists: "🗓️ Mon, Wed, Fri", "🗓️ Tuesday and Thursday"
- *
- * Notes:
- * - We only parse up to the next markup "<" to avoid pulling in HTML tags.
- * - Multiple calendar markers are supported; if any includes today, we return true.
- * - We accept both 🗓 and 🗓️ (with VS16) and common alternatives 📅, 📆
- */
-const DOW_NAMES: Record<string, number> = {
-	sun: 0,
-	sunday: 0,
-	mon: 1,
-	monday: 1,
-	tue: 2,
-	tues: 2,
-	tuesday: 2,
-	wed: 3,
-	weds: 3,
-	wednesday: 3,
-	thu: 4,
-	thur: 4,
-	thurs: 4,
-	thursday: 4,
-	fri: 5,
-	friday: 5,
-	sat: 6,
-	saturday: 6,
-};
 
-function normalizeWord(w: string): string {
-	return w.toLowerCase().replace(/[^a-z]/g, "");
-}
-
-function expandRange(start: number, end: number): number[] {
-	// Inclusive, circular week
-	const out: number[] = [];
-	let cur = start;
-	out.push(cur);
-	while (cur !== end) {
-		cur = (cur + 1) % 7;
-		out.push(cur);
-	}
-	return out;
-}
-
-function parseScheduleChunk(rawChunk: string): Set<number> {
-	const days = new Set<number>();
-	const raw = rawChunk.toLowerCase().trim();
-
-	// Simple buckets
-	if (/\bdaily\b|\bevery\s*day\b/.test(raw)) {
-		[0, 1, 2, 3, 4, 5, 6].forEach((d) => days.add(d));
-		return days;
-	}
-	if (
-		/\bweekdays?\b/.test(raw) ||
-		/\bmonday\s*[-–]\s*friday\b/.test(raw) ||
-		/\bmon\s*[-–]\s*fri\b/.test(raw)
-	) {
-		[1, 2, 3, 4, 5].forEach((d) => days.add(d));
-		return days;
-	}
-	if (
-		/\bweekends?\b/.test(raw) ||
-		/\bsat\s*[-–]\s*sun\b/.test(raw) ||
-		/\bsaturday\s*[-–]\s*sunday\b/.test(raw)
-	) {
-		[6, 0].forEach((d) => days.add(d));
-		return days;
-	}
-
-	// Ranges like "Mon–Fri", "Thu-Sun"
-	const rangeRe =
-		/(sun|sunday|mon|monday|tue|tues|tuesday|wed|weds|wednesday|thu|thur|thurs|thursday|fri|friday|sat|saturday)\s*[-–]\s*(sun|sunday|mon|monday|tue|tues|tuesday|wed|weds|wednesday|thu|thur|thurs|thursday|fri|friday|sat|saturday)/gi;
-	for (const m of raw.matchAll(rangeRe)) {
-		const a = DOW_NAMES[normalizeWord(m[1])];
-		const b = DOW_NAMES[normalizeWord(m[2])];
-		if (a !== undefined && b !== undefined) {
-			expandRange(a, b).forEach((d) => days.add(d));
-		}
-	}
-
-	// Expressions like "Every Sunday", "On Tuesday and Thursday"
-	const cleaned = raw
-		.replace(/\bevery\b/gi, " ")
-		.replace(/\bon\b/gi, " ")
-		.replace(/[:.;,]+/g, " ");
-
-	// Individual days list: split on commas/and/slashes/spaces and collect
-	const tokens = cleaned
-		.split(/[,/]|(?:\band\b)/gi)
-		.flatMap((t) => t.split(/\s+/))
-		.map((t) => normalizeWord(t.replace(/s\b/, ""))) // remove trailing plural 's'
-		.filter(Boolean);
-
-	for (const tok of tokens) {
-		const d = DOW_NAMES[tok];
-		if (d !== undefined) days.add(d);
-	}
-
-	return days;
-}
-
-/**
- * Returns true when the task text contains a recognized calendar schedule marker.
- * We accept 🗓, 🗓️, 📅, 📆.
- */
-export function hasCalendarScheduleMarker(task: TaskItem): boolean {
-	const txt = task?.text ?? "";
-	if (!txt) return false;
-	// u-flag for proper unicode handling
-	const markerRe = /(?:🗓️|🗓|📅|📆)/u;
-	return markerRe.test(txt);
-}
-
-export function isScheduledForToday(
-	task: TaskItem,
-	date = new Date()
-): boolean {
-	const txt = task?.text ?? "";
-	if (!txt) return false;
-
-	// Match any of the accepted calendar markers followed by the schedule text until markup/newline
-	const re = /(?:🗓️|🗓|📅|📆)\s*([^<\n\r]*)/giu;
-
-	const today = date.getDay(); // 0=Sun ... 6=Sat
-
-	for (const m of txt.matchAll(re)) {
-		const chunk = (m[1] ?? "").trim();
-		if (!chunk) continue;
-		const days = parseScheduleChunk(chunk);
-		if (days.has(today)) {
-			return true;
-		}
-	}
-	return false;
-}
